add: componentsに分解
