--- next-app/src/pages/cal_week/index.tsx
@@ -1,13 +1,0 @@
-import type { NextPage } from 'next'
-import CalWeek from 'src/components/calendar_week/CalWeek'
-
-const CalWeekPage: NextPage = () => {
-  
-  return(
-    <div>
-      <CalWeek/>
-    </div>
-  )
-}
-
-export default CalWeekPage
